Added all existing GCodeReader code
GCodeReader class has main code (takes in a text file of gcode, processes it, draws the image in a GUI, and prints out each command). Command class creates objects with a type, mode, and list of parameters. Params class creates objects that are pairs of the parameter type and its value. Line class creates line objects making storage of the info necessary to draw lines a little easier. The two text files are sample GCodes.